Fix breaking change.
