+++ lib/calculate-class-breaks/index.js
@@ -1,8 +1,8 @@
-const _ = require('lodash')
 const Classifier = require('classybrew')
 const normalizeClassificationValues = require('./normalize-classification-values')
 const calculateStdDevIntervals = require('./calculate-std-dev-intervals')
 const transformClassBreaksToRanges = require('./transform-class-breaks-to-ranges')
+const filterAndValidateClassificationFeatures = require('./filter-and-validate-classification-features')
 
 function calculateClassBreaks (features, classification) {
   if (!classification.method) throw new Error('must supply classification.method')
@@ -37,38 +37,14 @@ function transformValuesToClassBreaksArray (values, classification) {
 }
 
 function getFieldValues (features, classificationField) {
-  const values = features.map(feature => {
-    if (shouldSkipFeature({ feature, classificationField })) {
-      return
-    }
-
-    validateClassificationValue(feature, classificationField)
-
-    const value = feature.properties[classificationField]
-
-    if (value <= 0) {
-      return 0
-    }
-    return value
-  }).filter(value => {
-    return !_.isUndefined(value)
+  const values = filterAndValidateClassificationFeatures(features, classificationField).map(value => {
+    return value <= 0 ? 0 : value
   })
+
   if (!values || values.length === 0) {
     throw new Error(`"${classificationField}" was not found on any feature.`)
   }
   return values
-}
-
-function shouldSkipFeature ({ feature: { properties }, classificationField }) {
-  const value = properties[classificationField]
-  return value === undefined || value === null
-}
-
-function validateClassificationValue ({ properties }, classificationField) {
-  const value = properties[classificationField]
-  if (!_.isNumber(value)) {
-    throw new TypeError(`Cannot use non-numeric classificationField, ${classificationField}: ${value}`)
-  }
 }
 
 module.exports = calculateClassBreaks
